# API-CONTRACT.yaml
openapi: 3.1.0
info:
  title: Verdict API
  version: "1.0.0"
  description: >
    Canonical API contract for the Verdict hackathon judging platform.
    Every endpoint here must be implemented exactly as named — no additional
    endpoints invented by either the frontend or backend build without updating
    this file first. Enum values are defined once in ENUMS.md and referenced
    here by name. All list responses are wrapped in {"data": [...], "meta": {...}};
    all single-resource responses are wrapped in {"data": {...}}; all errors are
    {"error": {"code": "...", "message": "...", "details": {...}?}}.
servers:
  - url: /api/v1
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: verdict_session
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, enum: [unauthenticated, forbidden, not_found, validation_error, deadline_passed, event_not_open, conflict, rate_limited, internal_error] }
            message: { type: string }
            details: { type: object }
          required: [code, message]
      required: [error]
...
    ProjectRanking:
      type: object
      properties:
        project_id: { type: string, format: uuid }
        reviews_count: { type: integer }
        is_rankable: { type: boolean }
        official_score: { type: number, nullable: true }
        display_score: { type: number, nullable: true }
        rank: { type: integer, nullable: true }
        tie_break_reason: { type: string, nullable: true }
        flags: { type: array, items: { type: string } }
  parameters:
    eventSlug:
      name: event_slug
      in: path
      required: true
      schema: { type: string }
    projectId:
      name: project_id
      in: path
      required: true
      schema: { type: string, format: uuid }
security:
  - sessionCookie: []
paths:
...
  /events/{event_slug}:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get:
      summary: Event detail (public)
      security: []
      responses:
        '200': { description: ok }
        '404': { description: not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      summary: Update event settings (organizer/admin only)
      responses:
        '200': { description: ok }
        '403': { description: forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /events/{event_slug}/pairwise/next:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get: { summary: Get a pairwise comparison pair for the caller (judge), responses: { '200': { description: ok }, '204': { description: no pair available right now } } }
  /events/{event_slug}/pairwise:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    post:
      summary: Submit a pairwise comparison result
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                project_a_id: { type: string, format: uuid }
                project_b_id: { type: string, format: uuid }
                winner_project_id: { type: string, format: uuid }
              required: [project_a_id, project_b_id, winner_project_id]
      responses: { '201': { description: created }, '403': { description: forbidden } }

  # ---------------------------------------------------------------- ENGINE / RESULTS
  /events/{event_slug}/normalization/run:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    post:
      summary: Compute a new normalization run per JUDGING.md (organizer/admin)
      responses: { '201': { description: created }, '403': { description: forbidden } }
  /events/{event_slug}/results:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get:
      summary: Ranked results. Public only once events.status = results_published; organizer/admin can always view the latest run.
      security: []
      responses:
        '200': { description: ok, content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/ProjectRanking' } } } } } } }
        '403': { description: not yet published, and caller is not organizer/admin, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /events/{event_slug}/results/{project_id}/explain:
    parameters:
      - { $ref: '#/components/parameters/eventSlug' }
      - { $ref: '#/components/parameters/projectId' }
    get:
      summary: Full provenance chain per JUDGING.md §8, plus the separate informational panel
      responses: { '200': { description: ok }, '403': { description: not published and caller is not organizer/admin } }
  /events/{event_slug}/dashboard/progress:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get:
      summary: Judging Control Room data — coverage, per-judge completion, integrity flags (organizer/admin)
      responses: { '200': { description: ok }, '403': { description: forbidden } }

  # ---------------------------------------------------------------- VOTING
...